Convert Add-a-Job modal to a dedicated page
Replace the CreateJobModal on the jobs board with a full-page form at
/jobs/new. The "Add New Job" button becomes a router-link to the new
route, and the form's contents/validation/submission logic carry over
unchanged.

- Add CreateJobView.vue (page version of the form) and a /jobs/new route
  with SEO title/description
- Prerender /jobs/new for indexability, consistent with /recommend-event
- Cancel now navigates back to /jobs; success scrolls the window to top
- Remove the now-unused CreateJobModal.vue and update the JobsList test
  stub (create-job-modal -> router-link)
